+++ src/components/ads/ad/Ad.stories.ts
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/svelte-vite";
 import Ad from "./Ad.svelte";
 
-// More on how to set up stories at: https://storybook.js.org/docs/writing-stories
 const meta = {
   title: "Ads/Ad",
   component: "khao-ad",