+++ src/icons/types/IconName.ts
@@ -4,13 +4,17 @@ export const iconNames = [
   "copy",
   "external-link",
   "facebook",
+  "help",
+  "info",
   "instagram",
   "location",
   "newsletter",
+  "notification",
   "pinterest",
   "printer",
   "restaurant",
   "rss",
+  "speaker",
 ];
 
 export type IconName = (typeof iconNames)[number];