+++ src/icons/types/IconSizeFactor.ts
@@ -1,0 +1,18 @@
+export const iconSizeFactors = [
+  "4",
+  "5",
+  "6",
+  "7",
+  "8",
+  "9",
+  "10",
+  "11",
+  "12",
+  "13",
+  "14",
+  "15",
+];
+
+export type IconSizeFactor = (typeof iconSizeFactors)[number];
+
+export const iconSizeFactorDefault: IconSizeFactor = "4";