+++ CHANGELOG.md
@@ -5,9 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [NEXT]
+## [0.0.17]
 
-- added card component
+- added icon `help`
+
+## [0.0.16]
+
+- fix `khao-link` margin when there is no icon
+
+## [0.0.15]
+
+- added `khao-link` component
+- removed wrapper around svg icons in all components
+- added sizeFactor prop to icon component
+- added icon storys
+- increased size of `khao-map-marker-pin`
+
+## [0.0.14]
+
+- added title and icon to Card component
+- added `khao-card-notice` component
+
+## [0.0.13]
+
+- added `khao-card` component
 - brought components and their stories to the same folder
 - unified icon size for different button sizes
 
@@ -23,8 +44,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [0.0.10]
 
-- added khao-social-buttons
-- added khao-icon-button component
+- added `khao-social-buttons`
+- added `khao-icon-button` component
 - updated khao-malet to v0.0.7
 
 ## [0.0.0 - 0.0.9]